[FOLD] codec tidy
