+++ test/codec/deflate_service.cpp
@@ -8,23 +8,40 @@
 //
 
 // Test that header file is self-contained.
-#include <boost/http_proto/encoder.hpp>
+#include <boost/http_proto/codec/deflate_service.hpp>
 
+#include <boost/http_proto/context.hpp>
 #include "test_suite.hpp"
 
 namespace boost {
 namespace http_proto {
 
-class encoder_test
+class deflate_codec_test
 {
 public:
     void
+    testDecoder()
+    {
+        context ctx;
+        install_deflate_decoder(ctx);
+    }
+
+    void
+    testEncoder()
+    {
+        context ctx;
+        install_deflate_encoder(ctx);
+    }
+
+    void
     run()
     {
+        testDecoder();
+        testEncoder();
     }
 };
 
-TEST_SUITE(encoder_test, "boost.http_proto.encoder");
+TEST_SUITE(deflate_codec_test, "boost.http_proto.deflate_service");
 
 } // http_proto
 } // boost
